Comments about Question#6
This question has two parts and analysis is done separately.

The figure (“Income Category” vs “Total Transactions”) was created to show a relationship of “Income category” vs. “Total transaction”. As a first step, created a data frame with necessary number of columns. Found out that there is an “unknown” Income category which is approx. 10% of the whole data and was hard to predict that where it would fit in. So, decided to drop the rows with “unknown” Income Category and that makes the data clean to do the analysis. With income category as group by, determined the count of customers. Highest customer count is the one with lowest salary range and the lowest customer count is the one with highest salary range. It helps to determine that which income category customers are more likely to be the potential customers for the bank.
The figure (“Months Inactive” vs “Total Transactions”) was created to show a relationship of “Months Inactive” vs. “Total transaction”. As a first step, created a data frame with necessary number of columns. The graph shows that the highest number of transactions were made by the customers who were inactive for 3 months. The analysis just shows the trends of using the credit card. What we can predict here is the trend/frequency for the usage of credit card. We can see that most of the customers have a trend to use a card every three months on the average.